# refactor(batch): use memcomparable encoding to accelerate top-n executor