--- a/src/batch/Cargo.toml
+++ b/src/batch/Cargo.toml
@@ -86,3 +86,6 @@ harness = false
 name = "order_by"
 harness = false
 
+[[bench]]
+name = "top_n"
+harness = false