--- a/src/batch/src/executor/top_n.rs
+++ b/src/batch/src/executor/top_n.rs
@@ -12,17 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::cmp::Reverse;
 use std::collections::BinaryHeap;
-use std::sync::Arc;
 use std::vec::Vec;
 
 use futures_async_stream::try_stream;
-use risingwave_common::array::DataChunk;
+use itertools::Itertools;
+use risingwave_common::array::{DataChunk, Row};
 use risingwave_common::catalog::Schema;
 use risingwave_common::error::{Result, RwError};
-use risingwave_common::util::chunk_coalesce::DEFAULT_CHUNK_BUFFER_SIZE;
-use risingwave_common::util::sort_util::{HeapElem, OrderPair};
+use risingwave_common::util::chunk_coalesce::DataChunkBuilder;
+use risingwave_common::util::encoding_for_comparison::encode_chunk;
+use risingwave_common::util::sort_util::OrderPair;
 use risingwave_pb::batch_plan::plan_node::NodeBody;
 
 use crate::executor::{
@@ -30,69 +30,16 @@ use crate::executor::{
 };
 use crate::task::BatchTaskContext;
 
-struct TopNHeap {
-    order_pairs: Arc<Vec<OrderPair>>,
-    min_heap: BinaryHeap<Reverse<HeapElem>>,
-    size: usize,
-}
-
-impl TopNHeap {
-    fn insert(&mut self, elem: HeapElem) {
-        if self.min_heap.len() < self.size {
-            self.min_heap.push(Reverse(elem));
-        } else if elem > self.min_heap.peek().unwrap().0 {
-            self.min_heap.push(Reverse(elem));
-            self.min_heap.pop();
-        }
-    }
-
-    pub fn fit(&mut self, chunk: DataChunk) {
-        DataChunk::rechunk(&[chunk], 1)
-            .unwrap()
-            .into_iter()
-            .for_each(|c| {
-                let elem = HeapElem {
-                    order_pairs: self.order_pairs.clone(),
-                    chunk: c,
-                    chunk_idx: 0usize, // useless
-                    elem_idx: 0usize,
-                    encoded_chunk: None,
-                };
-                self.insert(elem);
-            });
-    }
-
-    pub fn dump(&mut self, offset: usize) -> Option<DataChunk> {
-        if self.min_heap.is_empty() {
-            return None;
-        }
-        let mut chunks = self
-            .min_heap
-            .drain_sorted()
-            .map(|e| e.0.chunk)
-            .collect::<Vec<_>>();
-        chunks.reverse();
-
-        if offset >= chunks.len() {
-            return None;
-        }
-
-        // Skip the first `offset` elements
-        if let Ok(mut res) = DataChunk::rechunk(&chunks[offset..], self.size - offset) {
-            assert_eq!(res.len(), 1);
-            Some(res.remove(0))
-        } else {
-            None
-        }
-    }
-}
-
+/// Top-N Executor
+///
+/// Use a N-heap to store the smallest N rows.
 pub struct TopNExecutor {
     child: BoxedExecutor,
-    top_n_heap: TopNHeap,
-    identity: String,
-    chunk_size: usize,
+    order_pairs: Vec<OrderPair>,
     offset: usize,
+    limit: usize,
+    schema: Schema,
+    identity: String,
 }
 
 #[async_trait::async_trait]
@@ -117,40 +64,36 @@ impl BoxedExecutorBuilder for TopNExecutor {
         Ok(Box::new(Self::new(
             inputs.remove(0),
             order_pairs,
-            top_n_node.get_limit() as usize,
             top_n_node.get_offset() as usize,
+            top_n_node.get_limit() as usize,
             source.plan_node().get_identity().clone(),
-            DEFAULT_CHUNK_BUFFER_SIZE,
         )))
     }
 }
 
 impl TopNExecutor {
-    fn new(
+    pub fn new(
         child: BoxedExecutor,
         order_pairs: Vec<OrderPair>,
-        limit: usize,
         offset: usize,
+        limit: usize,
         identity: String,
-        chunk_size: usize,
     ) -> Self {
+        let schema = child.schema().clone();
         Self {
-            top_n_heap: TopNHeap {
-                min_heap: BinaryHeap::new(),
-                size: limit + offset,
-                order_pairs: Arc::new(order_pairs),
-            },
             child,
-            identity,
-            chunk_size,
+            order_pairs,
             offset,
+            limit,
+            schema,
+            identity,
         }
     }
 }
 
 impl Executor for TopNExecutor {
     fn schema(&self) -> &Schema {
-        self.child.schema()
+        &self.schema
     }
 
     fn identity(&self) -> &str {
@@ -162,20 +105,70 @@ impl Executor for TopNExecutor {
     }
 }
 
+struct HeapElem {
+    encoded_row: Vec<u8>,
+    row: Row,
+}
+
+impl PartialEq for HeapElem {
+    fn eq(&self, other: &Self) -> bool {
+        self.encoded_row.eq(&other.encoded_row)
+    }
+}
+
+impl Eq for HeapElem {}
+
+impl PartialOrd for HeapElem {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        self.encoded_row.partial_cmp(&other.encoded_row)
+    }
+}
+
+impl Ord for HeapElem {
+    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+        self.encoded_row.cmp(&other.encoded_row)
+    }
+}
+
 impl TopNExecutor {
     #[try_stream(boxed, ok = DataChunk, error = RwError)]
-    async fn do_execute(mut self: Box<Self>) {
+    async fn do_execute(self: Box<Self>) {
+        let mut heap: BinaryHeap<HeapElem> = BinaryHeap::new();
+        let heap_size = self.offset + self.limit;
+
         #[for_await]
-        for data_chunk in self.child.execute() {
-            let data_chunk = data_chunk?;
-            self.top_n_heap.fit(data_chunk);
+        for chunk in self.child.execute() {
+            let chunk = chunk?.compact()?;
+            for (row_id, encoded_row) in encode_chunk(&chunk, &self.order_pairs)
+                .into_iter()
+                .enumerate()
+            {
+                if heap.len() < heap_size {
+                    let row = chunk.row_at_unchecked_vis(row_id).to_owned_row();
+                    heap.push(HeapElem { encoded_row, row });
+                }
+                // handle the case `heap_size == 0` 
+                else if let Some(peek) = heap.peek() && encoded_row < peek.encoded_row {
+                    let row = chunk.row_at_unchecked_vis(row_id).to_owned_row();
+                    heap.push(HeapElem { encoded_row, row });
+                    heap.pop();
+                }
+            }
         }
 
-        if let Some(data_chunk) = self.top_n_heap.dump(self.offset) {
-            let batch_chunks = DataChunk::rechunk(&[data_chunk], DEFAULT_CHUNK_BUFFER_SIZE)?;
-            for ret_chunk in batch_chunks {
-                yield ret_chunk
+        let mut chunk_builder = DataChunkBuilder::with_default_size(self.schema.data_types());
+        for HeapElem { row, .. } in heap
+            .drain()
+            .sorted_unstable()
+            .skip(self.offset)
+            .take(self.limit)
+        {
+            if let Some(spilled) = chunk_builder.append_one_row_from_datums(row.values())? {
+                yield spilled
             }
+        }
+        if let Some(spilled) = chunk_builder.consume_all()? {
+            yield spilled
         }
     }
 }
@@ -223,10 +216,9 @@ mod tests {
         let top_n_executor = Box::new(TopNExecutor::new(
             Box::new(mock_executor),
             order_pairs,
-            3,
             1,
-            "TopNExecutor2".to_string(),
-            DEFAULT_CHUNK_BUFFER_SIZE,
+            3,
+            "TopNExecutor".to_string(),
         ));
         let fields = &top_n_executor.schema().fields;
         assert_eq!(fields[0].data_type, DataType::Int32);
@@ -246,6 +238,50 @@ mod tests {
         }
 
         let res = stream.next().await;
+        assert!(matches!(res, None));
+    }
+
+    #[tokio::test]
+    async fn test_limit_0() {
+        let schema = Schema {
+            fields: vec![
+                Field::unnamed(DataType::Int32),
+                Field::unnamed(DataType::Int32),
+            ],
+        };
+        let mut mock_executor = MockExecutor::new(schema);
+        mock_executor.add(DataChunk::from_pretty(
+            "i i
+             1 5
+             2 4
+             3 3
+             4 2
+             5 1",
+        ));
+        let order_pairs = vec![
+            OrderPair {
+                column_idx: 1,
+                order_type: OrderType::Ascending,
+            },
+            OrderPair {
+                column_idx: 0,
+                order_type: OrderType::Ascending,
+            },
+        ];
+        let top_n_executor = Box::new(TopNExecutor::new(
+            Box::new(mock_executor),
+            order_pairs,
+            1,
+            0,
+            "TopNExecutor".to_string(),
+        ));
+        let fields = &top_n_executor.schema().fields;
+        assert_eq!(fields[0].data_type, DataType::Int32);
+        assert_eq!(fields[1].data_type, DataType::Int32);
+
+        let mut stream = top_n_executor.execute();
+        let res = stream.next().await;
+
         assert!(matches!(res, None));
     }
 }