I hereby agree to the terms of the [Singularity Data, Inc. Contributor License Agreement](https://gist.github.com/skyzh/0663682a70b0edde7ae991492f2314cb#file-s9y_cla).

## What's changed and what's your intention?

As title.

(BTW, if `limit` is set to 0, the compute node will panic. I fix the bug in this PR

## Checklist

- [x] I have written necessary rustdoc comments
- [x] I have added necessary unit tests and integration tests
- [x] All checks passed in `./risedev check` (or alias, `./risedev c`)

## Microbenchmark

*Row count: 1024 * 1024. Columns: (Int64). Offset: 128. Limit: 128.*
| Batch size | Before | After |
| --- | --- | --- |
| 32 | 310.02 ms |**91.436 ms**|
| 128 |323.96 ms|**87.710 ms** |
| 512 | 339.08 ms|**91.139 ms**|
| 1024 |338.54 ms |**90.840 ms**|
| 2048 |339.43 ms |**91.640 ms**|

*Row count: 1024 * 1024. Columns: (Int64, Varchar, Float32, Timestamp). Offset: 128. Limit: 128.*
| Batch size | Before | After |
| --- | --- | --- |
| 32 |1.1475 s |**348.22 ms**|
| 128 |1.2173 s |**338.44 ms**|
| 512 |1.2302 s |**364.34 ms**|
| 1024 |1.2275 s |**363.08 ms**|
| 2048 |1.2286 s |**371.52 ms**|

## Documentation

If your pull request contains user-facing changes, please specify the types of the changes, and create a release note. Otherwise, please feel free to remove this section.

### Types of user-facing changes

Please keep the types that apply to your changes, and remove those that do not apply.

* Installation and deployment 
* Connector (sources & sinks)
* SQL commands, functions, and operators
* RisingWave cluster configuration changes
* Other (please specify in the release note below)

### Release note

Please create a release note for your changes. In the release note, focus on the impact on users, and mention the environment or conditions where the impact may occur.

## Refer to a related PR or issue link (optional)